pyproject.toml
[project]
name = "grillcheese-amd-mvp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "fastapi>=0.127.0",
    "llama-cpp-python>=0.3.16",
    "numpy<2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "scikit-learn>=1.3.0",
    "sentence-transformers>=2.2.0",
    "torch==2.2.2",
    "torchvision>=0.17.0",
    "transformers>=4.57.3",
    "uvicorn>=0.40.0",
    "vulkan>=1.3.275.1",
    "websockets>=15.0.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
